requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
spotipy
PyYAML==6.0